+++ app/src/main/java/com/battlelancer/seriesguide/dataliberation/JsonExportTask.kt
@@ -466,6 +466,12 @@ open class JsonExportTask(
         const val TMDB_SHOW = "tmdb-show"
         const val SEASON = "season"
         const val EPISODE = "episode"
+
+        /**
+         * Special type only supported when importing and movie with that IMDB ID is in the
+         * database already.
+         */
+        const val IMDB_MOVIE = "imdb-movie"
     }
 
 }