+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Releases marked with 🧪 (or previously with the "beta" suffix) were released o
 ### Next release
 
 * 🔧 Import: support importing movies with only IMDB IDs.
+* 🔧 Import: support new "imdb-movie" list item type, maps to added movie using "externalId" as IMDB ID on import, otherwise skips.
 * 🔧 Import: restore original data on failure.
 
 ## Version 2026.2